third_party/WebKit/LayoutTests/media/track/track-selection-metadata.html

<!DOCTYPE html>
<title>Multiple 'metadata' tracks with 'default'</title>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<video src="../content/test.ogv">
  <track kind="metadata" src="captions-webvtt/tc005-default-styles.vtt" id="t1">
  <track kind="metadata" src="captions-webvtt/tc025-class.vtt" default id="t2hidden">
  <track kind="metadata" src="captions-webvtt/tc005-metadata-area.vtt" id="t3">
  <track kind="metadata" src="captions-webvtt/tc004-webvtt-file.vtt" default id="t4hidden">
</video>
<script>
async_test(function() {
    var video = document.querySelector('video');
    video.onloadstart = this.step_func(function() {
        var tracks = video.textTracks;
        assert_equals(tracks.length, 4);
        for (var i = 0; i < tracks.length; ++i) {
            assert_equals(tracks[i].kind, 'metadata');

            var trackElement = document.getElementById(tracks[i].id);
            if (tracks[i].id.indexOf('hidden') != -1) {
                assert_true(trackElement.default);
                assert_equals(tracks[i].mode, 'hidden');
            } else {
                assert_false(trackElement.default);
                assert_equals(tracks[i].mode, 'disabled');
            }
        }
        this.done();
   });
});
</script>
